Add low-pass filter and smooth box rendering

+++ src/Writer/NcursesWriter.h
@@ -34,7 +34,7 @@ class NcursesWriter
 
     virtual ~NcursesWriter();
 
-    virtual void write(int32_t row, int32_t column, ColorDefinition color,
+    virtual void write(double row, double column, ColorDefinition color,
                        const std::wstring &msg, wchar_t character);
 
     virtual void clear();
